refactor pm quality score run lookups
