+++ src/app/router_registry.py
@@ -85,6 +85,9 @@ from app.routers.dpm_command_center_pm_quality_review_action_lookup import (
 from app.routers.dpm_command_center_pm_quality_review_actions import (
     router as dpm_command_center_pm_quality_review_actions_router,
 )
+from app.routers.dpm_command_center_pm_quality_score_run_lookup import (
+    router as dpm_command_center_pm_quality_score_run_lookup_router,
+)
 from app.routers.dpm_command_center_pm_quality_summary_invocations import (
     router as dpm_command_center_pm_quality_summary_invocations_router,
 )
@@ -282,6 +285,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_pm_quality_policies_router)
     app.include_router(dpm_command_center_pm_quality_review_actions_router)
     app.include_router(dpm_command_center_pm_quality_review_action_lookup_router)
+    app.include_router(dpm_command_center_pm_quality_score_run_lookup_router)
     app.include_router(dpm_command_center_pm_quality_summary_invocations_router)
     app.include_router(dpm_command_center_pm_quality_summary_lookup_router)
     app.include_router(dpm_command_center_portfolio_memory_router)
